LocalCompositionProvider.kt is not providing the activity.
