+++ mifos-passcode-cmp/src/androidMain/kotlin/auth/deviceAuth/PlatformAuthenticator.kt
@@ -1,0 +1,121 @@
+package com.mifos.passcode.auth.deviceAuth
+
+import android.app.KeyguardManager
+import android.content.Context
+import android.content.Intent
+import android.os.Build
+import android.provider.Settings
+import androidx.annotation.RequiresApi
+import androidx.biometric.BiometricManager
+import androidx.biometric.BiometricManager.Authenticators.BIOMETRIC_STRONG
+import androidx.biometric.BiometricManager.Authenticators.DEVICE_CREDENTIAL
+import androidx.biometric.BiometricPrompt
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.fragment.app.FragmentActivity
+import com.mifos.passcode.auth.deviceAuth.domain.AuthenticationResult
+import com.mifos.passcode.biometric.domain.AuthenticatorStatus
+import kotlinx.coroutines.suspendCancellableCoroutine
+import kotlin.coroutines.resume
+
+
+actual class PlatformAuthenticator private actual constructor() {
+
+    private var applicationContext: FragmentActivity? = null
+
+    actual constructor(activity: Any?) : this() {
+        this@PlatformAuthenticator.applicationContext = activity as? FragmentActivity
+    }
+
+    private val bioMetricManager = applicationContext?.let {
+        BiometricManager.from(it)
+    }
+    private val authenticatorStatus by  mutableStateOf(AuthenticatorStatus())
+
+    actual fun getDeviceAuthenticatorStatus(): AuthenticatorStatus {
+
+        val result = bioMetricManager?.canAuthenticate(BIOMETRIC_STRONG) ?: 1
+
+        when(result){
+            BiometricManager.BIOMETRIC_ERROR_HW_UNAVAILABLE -> {
+                authenticatorStatus.message = "Hardware unavailable. Try again later."
+            }
+
+            BiometricManager.BIOMETRIC_ERROR_NONE_ENROLLED -> {
+                authenticatorStatus.message ="Biometrics not enrolled."
+                authenticatorStatus.biometricsNotPossible = false
+            }
+
+            BiometricManager.BIOMETRIC_ERROR_NO_HARDWARE -> {
+                authenticatorStatus.message = "Biometrics not available."
+            }
+
+            BiometricManager.BIOMETRIC_ERROR_SECURITY_UPDATE_REQUIRED -> {
+                authenticatorStatus.message = "Vulnerabilities found. Security update required."
+            }
+
+            BiometricManager.BIOMETRIC_ERROR_UNSUPPORTED -> {
+                authenticatorStatus.message = "Android version not supported."
+            }
+
+            BiometricManager.BIOMETRIC_STATUS_UNKNOWN -> {
+                authenticatorStatus.message = "Unable to determine whether the user can authenticate."
+            }
+
+            BiometricManager.BIOMETRIC_SUCCESS -> {
+                authenticatorStatus.biometricsSet = true
+                authenticatorStatus.biometricsNotPossible = false
+                authenticatorStatus.message = "Biometrics are set."
+            }
+        }
+
+        return authenticatorStatus
+    }
+
+    @RequiresApi(Build.VERSION_CODES.R)
+    actual fun setDeviceAuthOption() {
+        val enrollBiometric = Intent(Settings.ACTION_BIOMETRIC_ENROLL).apply {
+            putExtra(
+                Settings.EXTRA_BIOMETRIC_AUTHENTICATORS_ALLOWED,
+                BIOMETRIC_STRONG or DEVICE_CREDENTIAL
+            )
+        }
+        this.applicationContext?.startActivity(enrollBiometric)
+    }
+
+    actual suspend fun authenticate(title: String): AuthenticationResult = suspendCancellableCoroutine{ continuation ->
+
+        val promptInfo = BiometricPrompt.PromptInfo.Builder()
+            .setTitle(title)
+            .setSubtitle("Unlock using your PIN, Password, Pattern, Face or Fingerprint")
+            .setAllowedAuthenticators(BIOMETRIC_STRONG or DEVICE_CREDENTIAL)
+            .build()
+
+        applicationContext?.let {fragmentActivity ->
+            val prompt = BiometricPrompt(
+                fragmentActivity,
+                object: BiometricPrompt.AuthenticationCallback(){
+
+                    override fun onAuthenticationError(errorCode: Int, errString: CharSequence) {
+                        super.onAuthenticationError(errorCode, errString)
+                        continuation.resume(AuthenticationResult.Error("${errorCode}: $errString"))
+                    }
+
+                    override fun onAuthenticationFailed() {
+                        super.onAuthenticationFailed()
+                        AuthenticationResult.Failed()
+                    }
+
+                    override fun onAuthenticationSucceeded(result: BiometricPrompt.AuthenticationResult) {
+                        super.onAuthenticationSucceeded(result)
+                        continuation.resume(AuthenticationResult.Success())
+                    }
+                }
+            )
+
+            prompt.authenticate(promptInfo)
+        }
+
+    }
+
+}
