+++ mifos-passcode-cmp/src/commonMain/kotlin/auth/deviceAuth/PlatformAuthenticator.kt
@@ -1,14 +1,15 @@
-package com.mifos.passcode.auth.deviceAuth.domain
+package com.mifos.passcode.auth.deviceAuth
 
+import com.mifos.passcode.auth.deviceAuth.domain.AuthenticationResult
 import com.mifos.passcode.biometric.domain.AuthenticatorStatus
 
+expect class PlatformAuthenticator private constructor(){
 
-interface PlatformAuthenticator {
-
+    constructor(activity: Any?)
     fun getDeviceAuthenticatorStatus(): AuthenticatorStatus
 
     fun setDeviceAuthOption()
 
     suspend fun authenticate(title: String = ""): AuthenticationResult
-}
 
+}
