+++ mifos-passcode-cmp/src/commonMain/kotlin/ui/screen/DeviceAuthScreen.kt
@@ -7,6 +7,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.saveable.rememberSaveable
@@ -15,8 +16,11 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import com.mifos.passcode.LocalAndroidActivity
+import com.mifos.passcode.LocalPlatformAuthenticator
 import com.mifos.passcode.auth.AuthOption
 import com.mifos.passcode.auth.PlatformAuthOptions
+import com.mifos.passcode.auth.deviceAuth.PlatformAuthenticator
 import com.mifos.passcode.auth.deviceAuth.domain.AuthenticationResult
 import com.mifos.passcode.getPlatform
 import com.mifos.passcode.ui.component.MifosIcon
@@ -25,6 +29,9 @@ import com.mifos.passcode.ui.components.SystemAuthenticatorButton
 import com.mifos.passcode.ui.viewmodels.DeviceAuthenticatorViewModel
 import io.github.openmf.mifos_passcode_cmp.generated.resources.Res
 import io.github.openmf.mifos_passcode_cmp.generated.resources.app_name
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.coroutineScope
+import kotlinx.coroutines.flow.stateIn
 import org.jetbrains.compose.resources.stringResource
 
 
@@ -32,23 +39,22 @@ import org.jetbrains.compose.resources.stringResource
 fun DeviceAuthScreen(
     authOption: AuthOption? = null,
     onDeviceAuthSuccess: () -> Unit = {},
-    deviceAuthenticatorViewModel: DeviceAuthenticatorViewModel,
 ) {
 
+    val deviceAuthenticatorViewModel = DeviceAuthenticatorViewModel(LocalPlatformAuthenticator.current)
 
-    val authenticationResult =
-        deviceAuthenticatorViewModel.authenticationResult.collectAsStateWithLifecycle()
+    val authenticationResult = deviceAuthenticatorViewModel.authenticationResult.collectAsStateWithLifecycle()
 
-
-    val authenticatorStatus =
-        deviceAuthenticatorViewModel.authenticatorStatus.collectAsStateWithLifecycle()
+    val authenticatorStatus = deviceAuthenticatorViewModel.authenticatorStatus.collectAsStateWithLifecycle()
 
 
     var showAuthPrompt by rememberSaveable() {
         mutableStateOf(false)
     }
 
-    var showSetBiometricDialog by rememberSaveable(){ mutableStateOf(false) }
+    var showSetBiometricDialog by rememberSaveable{
+        mutableStateOf(false)
+    }
 
 
     if(authenticationResult.value == AuthenticationResult.Success() && showAuthPrompt){
