+++ sample/composeApp/src/androidMain/kotlin/com/mifos/passcode/sample/di/Module.android.kt
@@ -1,7 +1,7 @@
 package com.mifos.passcode.sample.di
 
-import com.mifos.passcode.auth.AuthOption
 import auth.AuthOptionAndroid
+import com.mifos.passcode.auth.AuthOption
 import org.koin.core.module.Module
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.bind
@@ -9,6 +9,5 @@ import org.koin.dsl.module
 
 
 actual val platformModule: Module = module {
-
     singleOf(::AuthOptionAndroid).bind<AuthOption>()
 }
