+++ sample/composeApp/src/commonMain/kotlin/com/mifos/passcode/sample/App.kt
@@ -2,6 +2,8 @@
 
 import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
+import com.mifos.passcode.LocalAndroidActivity
+import com.mifos.passcode.LocalCompositionProvider
 import com.mifos.passcode.sample.navigation.PasscodeNavigation
 import org.jetbrains.compose.ui.tooling.preview.Preview
